Split clause helpers into separate files
