+++ frontend/src/metabase-lib/v1/expressions/diagnostics/expression/check-arg-validators.ts
@@ -1,6 +1,6 @@
 import * as Lib from "metabase-lib";
 
-import { getClauseDefinition } from "../../config";
+import { getClauseDefinition } from "../../clause";
 import { visit } from "../../visitor";
 import { error } from "../utils";
