+++ frontend/src/metabase-lib/v1/expressions/diagnostics/expression/check-known-functions.ts
@@ -2,7 +2,7 @@ import { t } from "ttag";
 
 import * as Lib from "metabase-lib";
 
-import { getClauseDefinition } from "../../config";
+import { getClauseDefinition } from "../../clause";
 import { visit } from "../../visitor";
 import { error } from "../utils";
