+++ frontend/src/metabase-lib/v1/expressions/diagnostics/expression/check-supported-functions.ts
@@ -3,7 +3,7 @@ import { t } from "ttag";
 import * as Lib from "metabase-lib";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
 
-import { getClauseDefinition } from "../../config";
+import { getClauseDefinition } from "../../clause";
 import { getDatabase } from "../../utils";
 import { visit } from "../../visitor";
 import { error } from "../utils";
