+++ frontend/src/metabase-lib/v1/expressions/index.ts
@@ -5,3 +5,4 @@ export * from "./types";
 export * from "./diagnostics";
 export * from "./position";
 export * from "./errors";
+export * from "./clause";
